Add files via upload
In this proposed work, we are envisioning a military wellbeing administration (MHS) stage which depends on progressive IoT design. We propose a semantic Edge based system display which assumes a noteworthy job for conveying strategic and non-strategic snippet of data over the system. Further, the trading of data and resulting information investigation on the MHS makes the framework keen and smart. The soldier Health and Position Tracking System enables military to follow the present GPS position of soldier and furthermore checks the wellbeing status including temperature of body and beats of soldier. The System will likewise comprises of an additional component with the assistance of that soldier can request help physically or send a trouble flag to military if he is in need. The GPS modem sends the latitude and longitude position with association plan with the help of that military can pursue the present position of the soldier. The framework might be exceptionally useful for getting wellbeing status data of soldier and giving them moment help. Information will be sent to cloud legitimately from where we can download as csv and can apply data mining for examination.
The IoT makes the whole checking procedure quick, effective and the choices can be taken in less measure of time. So by utilizing these types of gear we are attempting to execute the essential life guarding framework for soldier in minimal effort and high dependability